[Distance Sensor] Added Distance signal processing filter

+++ esp_sensor_ecu/components/sensor/include/DistanceSensor.h
@@ -5,7 +5,7 @@ typedef struct DistanceSensor DistanceSensor;
 
 typedef struct {
     uint64_t timestamp;
-    uint8_t distance;
+    uint16_t distance;
 } DistanceData;
 
 extern QueueHandle_t distance_source_queue;
